refactor: change extensions from .scss to .css
