+++ src/features/SideBar/ui/DescDeveloper.module.css
