+++ src/features/SideBar/ui/DescDeveloper1.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from 'next-i18next';
 
-import cls from '@/features/SideBar/ui/DescDeveloper.module.scss';
+import cls from '@/features/SideBar/ui/DescDeveloper.module.css';
 
 export const DescDeveloper1 = () => {
   const { t } = useTranslation('common');
