+++ src/shared/ui/LangSwitcher/LangSwitcher.module.css
@@ -11,14 +11,14 @@
 }
 
 .button {
-  min-width: 40px;
+  min-width: 50px;
   height: 36.5px;
   padding: 10px;
 }
 
 .button:disabled {
   pointer-events: none;
-  min-width: 40px;
+  min-width: 50px;
   color: white;
   background-color: #4e4e4e;
 }
